Add warning when predict has a type mismatch
As discussed in (#731), when predict's `newdata` has a different dtype
in the FEs, the function will yield missing values only, this because
predict relies on a dictionary of FEs where keys are transformed to
strings and will fail to get the correct value (i.e. 10 vs "10.0").

This commit adds a warning in the `predict` method and a simple unit
test that checks for this case for the user.