#890 - Show how to replicate Stata results
.qmd file showing how to replicate common Stata results in PyFixest.

Note: I ran each part in both Stata and PyFixest and got the same results everywhere except for `fit4` where I show how to get HC3 standard errors.  For some reason the HC3 errors in PyFixest and Stata do not match.  I haven't had a chance to try to figure out why though.

If you know why, let me know and I can update the example.  If not, I can try to figure it out at a later date.  In the meantime we can remove that section or add a note that the HC3 standard errors do not match.